feat: activity ssr (#2965)
