+++ src/app/(activity)/(event)/events/page.tsx
@@ -5,7 +5,7 @@ import { useSuspenseInfiniteQuery } from '@tanstack/react-query';
 import { compact } from 'lodash-es';
 
 import { getActivityListItem } from '@/components/Activity/ActivityListItem.js';
-import { NavigationBar } from '@/components/Activity/NavigationBar.js';
+import { ActivityNavigationBar } from '@/components/Activity/ActivityNavigationBar.js';
 import { ListInPage } from '@/components/ListInPage.js';
 import { ScrollListKey, Source } from '@/constants/enum.js';
 import { createIndicator } from '@/helpers/pageable.js';
@@ -29,9 +29,9 @@ export default function Page() {
     return (
         <div className="flex w-full flex-col">
             {fireflyBridgeProvider.supported ? (
-                <NavigationBar>
+                <ActivityNavigationBar>
                     <Trans>Exclusive Events</Trans>
-                </NavigationBar>
+                </ActivityNavigationBar>
             ) : (
                 <div className="sticky top-0 z-20 hidden h-12 w-full flex-row items-center bg-primaryBottom px-4 text-xl font-bold md:flex">
                     <span>
